pyproject.toml
[tool.poetry]
name = "dnd5e-reddit-scraper"
version = "0.1.0"
description = ""
authors = ["Beaver <8586904+PragmaticBeaver@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "dnd5e_reddit_scraper"}]

[tool.poetry.dependencies]
python = "^3.10.9"
praw = "^7.6.1"
python-dotenv = "^0.21.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
